pyproject.toml
[tool.poetry]
name = "commerce"
version = "0.1.0"
description = ""
authors = ["rodvicj <rodvicj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
fontawesomefree = "^6.4.2"
# channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework = "^3.14.0"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true
